+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## dev
 
+* Fixes handling naive datetimes in the admin's requests overview.