+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## dev
 
 * Fixes handling naive datetimes in the admin's requests overview.
+* Fixed url() deprecation warnings for Django >= 3.1.