+++ CHANGELOG.md
@@ -8,3 +8,5 @@
 * Update isort and flake8 rules, reformat code with recent black (21.5b) and update code style.
 * Remove deprecated django.utils.http.quote import
 * Switch Request.id to BigAutoField
+* Support for Python 3.6, 3.7, 3.8, 3.9.
+* Support for Django >= 2.2.